couple small style improvements
