+++ packages/core/src/events/EventContext.ts
@@ -3,6 +3,4 @@ import { EventHandlers } from "./EventHandlers";
 
 export const EventHandlerContext = createContext<EventHandlers>(null);
 
-export const useEventHandler = () => {
-  return useContext(EventHandlerContext);
-};
+export const useEventHandler = () => useContext(EventHandlerContext);
